Put the finishing touches to the touchpadlib.
commit 1abbe0e4175dc32984d3c64b0b129b7423368784
Author: Mateusz Piotrowski <mpp302@gmail.com>
Date:   Wed Apr 20 22:49:52 2016 +0200

    Document touchpadlib.c and improve formatting.

    Improve error and warning messages.
    Issue #45.

commit d1c256dd388aa8521589ee7897eba0fcd851d447
Author: Mateusz Piotrowski <mpp302@gmail.com>
Date:   Wed Apr 20 22:41:52 2016 +0200

    Fix a bug with converting an Python 1 to c_int.

    I don't know why it is not a good idea but it does not work.
    Hence I remove it.

commit cc07a398689bd6d60e280f0e97392e2f0e2c0be1
Merge: b758371 492e956
Author: Mateusz Piotrowski <mpp302@gmail.com>
Date:   Wed Apr 20 02:04:22 2016 +0200

    Merge branch 'master' into detect-abs-xy-values

commit 492e95609701e506b8faf7396afe73d5b948f437
Author: Michal <michal7352@gmail.com>
Date:   Tue Apr 19 23:11:53 2016 +0200

    y-coordinate read fix

+++ app/touchpadlib/touchpadlib.py
@@ -155,7 +155,7 @@ class Touchpadlib:
         fetch_event = cls.touchpadlib.fetch_touchpad_event
         file_descriptor = cls.touchpad_file_descriptor
         event = cls.touchpad_event
-        if fetch_event(file_descriptor, event) == ctypes.c_int(1):
+        if fetch_event(file_descriptor, event) == 1:
             print("ERROR: touchpadlib fetch error.")
             cls.interrupt_and_finish()
